Added L-System generation (rules can be found in Tree.java), added flags and key presses to change between regular trees and fractal-trees, increased road offset so it doesn't clip when viewed at a distance and at a sharp angle.
For L-System:
Standard no. of iterations is 3
Press L to toggle between regular tree and fractal tree
Press J to decrease no. of iterations
Press K to increase no. of iterations
WARNING: As iterations increases, it will lag a lot. (Try not to go over 5)
The fractal trees will take the same spots as the regular trees as specified in the Level file
The trees themselves don't look all that good, but it's something.